% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cont_boxplot.R
\name{cont_boxplot_app}
\alias{cont_boxplot_app}
\title{Explore the relationship between continous and categorical
data}
\usage{
cont_boxplot_app(dataset, height = NULL)
}
\arguments{
\item{dataset}{}
}
\value{

}
\description{
Explore the relationship between continous and categorical
data
}
\examples{
library(ggplot2)
data(diamonds)
cont_boxplot_app(diamonds)

}
