% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cont_histogram.R
\name{cont_hist_app}
\alias{cont_hist_app}
\title{App for exploring data using histograms}
\usage{
cont_hist_app(dataset)
}
\arguments{
\item{dataset}{\itemize{
\item dataset that contains numeric variables
}}
}
\value{
Shiny App
}
\description{
This is an embeddable app that lets you explore the distribution
of continuous variables
}
\details{
This can be embedded into rmarkdown documents that use
runtime: shiny.
}
\examples{

library(ggplot2)
data(diamonds)
cont_hist_app(diamonds)

}
