% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cont_scatter.R
\name{cont_scatter_app}
\alias{cont_scatter_app}
\title{App for exploring data using scatter plots}
\usage{
cont_scatter_app(dataset)
}
\arguments{
\item{dataset}{}
}
\value{
shiny app that can be embedded in a doc
}
\description{
This is an embeddable app that lets you explore the relationship
between continuous variables.
}
\details{
This can be embedded into rmarkdown documents that use
runtime: shiny.
}
\examples{

library(ggplot2)
data(diamonds)
cont_scatter_app(diamonds)
}
