% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_dashboard.R
\name{explore_data2}
\alias{explore_data2}
\title{Explore data with a shiny app}
\usage{
explore_data2(dataset, data_dictionary = NULL)
}
\arguments{
\item{dataset}{\itemize{
\item dataset to explore
}}

\item{data_dictionary}{\itemize{
\item optional. Data dictionary with
information about variables
}}
}
\value{

}
\description{
Explore data with a shiny app
}
