% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_category_variables}
\alias{get_category_variables}
\title{Return a list of categories given a data.frame}
\usage{
get_category_variables(df)
}
\arguments{
\item{df}{}
}
\value{
named list with each entry listing all categories within a variable in the df
}
\description{
Return a list of categories given a data.frame
}
