% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_numeric_variables}
\alias{get_numeric_variables}
\title{Return the numeric categorical variables in a data frame}
\usage{
get_numeric_variables(df)
}
\arguments{
\item{df}{}
}
\value{
character vector of categorial variables in data.frame
}
\description{
Return the numeric categorical variables in a data frame
}
