% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xplor.R
\name{build_shiny_app}
\alias{build_shiny_app}
\title{Make a shiny app from a dataset}
\usage{
build_shiny_app(dataset, covariates, outcome_var, data_dictionary = NULL)
}
\arguments{
\item{dataset}{\itemize{
\item data.frame or data.table of data. This will get saved into \code{data/dataset.rds} for the app.
}}

\item{covariates}{\itemize{
\item character vector of covariates, corresponding to the column names of dataset
}}

\item{outcome_var}{\itemize{
\item character of variable name in dataset correpsonding to the outcome of interest
}}

\item{data_dictionary}{\itemize{
\item optional; a data frame that describes the variables in the dataset.
}

This function allows you to build a shiny app for a dataset. For more info, please see the
"Making a Burro App" vignette.}
}
\value{
shiny app in a folder
}
\description{
Make a shiny app from a dataset
}
\examples{
#create a new project/folder in RStudio before you run this
library(ggplot2)
data(diamonds)
covars <- colnames(diamonds)
if(interactive()){
burro::build_shiny_app(dataset=diamonds,
covariates=covars, outcome_var="cut")
}
}
