% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_data}
\alias{check_data}
\title{Checks the Dataset and sets attributes on it for use in}
\usage{
check_data(dataset, covariates = NULL, outcome_var = NULL)
}
\arguments{
\item{outcome_var}{}
}
\value{

}
\description{
Checks the Dataset and sets attributes on it for use in
}
