% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xplor.R
\name{explore_data}
\alias{explore_data}
\title{Explore a dataaset using a simple app}
\usage{
explore_data(dataset, covariates = NULL, outcome_var = NULL,
  data_dictionary = NULL)
}
\arguments{
\item{dataset}{\itemize{
\item a data frame
}}

\item{covariates}{\itemize{
\item Optional. a character vector of column names you want
to include in the dataset
}}

\item{outcome_var}{\itemize{
\item a categorical variable that describes an outcome to examine
}}

\item{data_dictionary}{\itemize{
\item Optional. A data.frame describing covariates. This is passed
on to DT:data.table() so it can be easily searched. There is no required format,
but the example uses two columns: \code{VariableName} and \code{Description}.
}

\code{explore_data} gives you a simple shiny app to explore data, The app is tailored to the
covariate types of the data.frame/data.table that you pass to it.

Each menu item covers a different part of exploratory data analysis with
simple visualizations for discussion.

The tabs are automatically updated with variables depending on the variable types.}
}
\value{
A Shiny App object that can either be run in the console or on an external
shiny server such as shinyapps.io. For building the actual app.R, see
build_shiny_app
}
\description{
Explore a dataaset using a simple app
}
\examples{
library(ggplot2)
data(diamonds)
if(interactive()){
   burro::explore_data(diamonds, outcome_var="cut")
}


# look at NHANES data (you need to have NHANES package installed)
library(NHANES)
data(NHANES)
if(interactive()){
  #this data dictionary is provided since NHANES doesn't have one
  data_dict <- read.csv(system.file("nhanes/data_dictionary.csv", package="burro"))
  burro::explore_data(NHANES[1:10,], data_dictionary=data_dict)
}
}
